+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2026-04-17
+
+### Added
+- Settings window accessible from the right-click menu
+- "Launch at login" toggle (uses SMAppService)
+- App info (name, version, icon) and Quit button in settings
+
 ## [0.3.0] - 2026-03-06
 
 ### Changed